Move `VersionLabel` to `app-sidebar.tsx`
